Adds very basic lighting

+++ Engine/Engine/Source/Private/MyWorld/Generation/ChunkGenerator.cpp
@@ -32,7 +32,7 @@ void Jafg::ChunkGenerator::ShapeChunk(const LSharedChunkArgs* SharedArgs, const 
     {
         for (LChunkKeyDomainTy Y = 0; Y < MwStatics::ChunkSize; Y++)
         {
-            const int32 Height = Maths::Floor(NoiseOutput[++Index] * 32) + 32;
+            const int32 Height = static_cast<int32>(Maths::Floor(NoiseOutput[++Index] * 32.0f) + 32.0f);
             for (LChunkKeyDomainTy Z = 0; Z < MwStatics::ChunkSize; Z++)
             {
                 const int32 MapZ = InKey.Z * MwStatics::ChunkSize + Z;
